+++ friends/src/app/queries/useFriendPrimaryTable.ts
@@ -1,5 +1,5 @@
 import { useQuery } from '@tanstack/react-query';
-import { useServiceContext } from '../../main';
+import { useServiceContext } from '../contexts/ServiceContext';
 import { FriendId } from '../../entities/friend';
 import { CourseBook } from '../../entities/courseBook';
