+++ friends/src/app/queries/useFriends.ts
@@ -1,5 +1,5 @@
 import { useQuery } from '@tanstack/react-query';
-import { useServiceContext } from '../../main';
+import { useServiceContext } from '../contexts/ServiceContext';
 
 export const useFriends = (req: { state: 'REQUESTED' | 'REQUESTING' | 'ACTIVE' }) => {
   const { friendService } = useServiceContext();
