+++ friends/src/app/screens/MainScreen/ManageFriendsDrawerContent/ManageFriendsDrawerContentRequestedList/index.tsx
@@ -2,7 +2,7 @@ import { FlatList, StyleSheet, Text, View } from 'react-native';
 import { useFriends } from '../../../../queries/useFriends';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { FriendId } from '../../../../../entities/friend';
-import { useServiceContext } from '../../../../../main';
+import { useServiceContext } from '../../../../contexts/ServiceContext';
 import { Button } from '../../../../components/Button';
 
 type Props = {};
