+++ friends/src/main.tsx
@@ -1,18 +1,15 @@
-import { createContext, useContext, useMemo } from 'react';
+import { useMemo } from 'react';
 import { ErrorBoundary } from './app/components/ErrorBoundary';
 import { Text, View } from 'react-native';
 import { createFetchClient } from './infrastructures/createFetchClient';
 import { App } from './app/App';
-import { TimetableViewService } from './usecases/timetableViewService';
 import { createTimetableViewService } from './infrastructures/createTimetableViewService';
-import { ColorService } from './usecases/colorService';
 import { createColorService } from './infrastructures/createColorService';
 import { createColorRepository } from './infrastructures/createColorRepository';
-import { FriendService } from './usecases/friendService';
 import { createFriendRepository } from './infrastructures/createFriendRepository';
 import { createFriendService } from './infrastructures/createFriendService';
 import { createCourseBookService } from './infrastructures/createCourseBookService';
-import { CourseBookService } from './usecases/courseBookService';
+import { serviceContext } from './app/contexts/ServiceContext';
 
 type ExternalProps = {
   'x-access-token': string;
@@ -24,18 +21,6 @@ type ExternalProps = {
   'x-device-id': string;
   'x-app-type': string;
   'x-app-version': string;
-};
-type ServiceContext = {
-  timetableViewService: TimetableViewService;
-  colorService: ColorService;
-  friendService: FriendService;
-  courseBookService: CourseBookService;
-};
-const serviceContext = createContext<ServiceContext | null>(null);
-export const useServiceContext = () => {
-  const context = useContext(serviceContext);
-  if (!context) throw new Error('context not provided');
-  return context;
 };
 
 export const Main = ({ 'x-access-token': xAccessToken, 'x-access-apikey': xAccessApikey }: ExternalProps) => {
